Move FAB to its own components
