+++ composeApp/src/commonMain/kotlin/com/github/familyvault/ui/screens/main/MainScreen.kt
@@ -5,19 +5,10 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.AddTask
 import androidx.compose.material.icons.filled.GroupAdd
-import androidx.compose.material.icons.filled.UploadFile
-import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import cafe.adriel.voyager.core.screen.Screen
 import cafe.adriel.voyager.navigator.LocalNavigator
@@ -26,23 +17,16 @@ import cafe.adriel.voyager.navigator.tab.CurrentTab
 import cafe.adriel.voyager.navigator.tab.LocalTabNavigator
 import cafe.adriel.voyager.navigator.tab.TabNavigator
 import com.github.familyvault.models.enums.chat.ChatThreadType
-import com.github.familyvault.services.IDocumentPickerService
-import com.github.familyvault.services.IFileCabinetService
-import com.github.familyvault.services.IImagePickerService
 import com.github.familyvault.states.ITaskListState
-import com.github.familyvault.ui.components.dialogs.CircularProgressIndicatorDialog
 import com.github.familyvault.ui.components.overrides.NavigationBar
 import com.github.familyvault.ui.screens.main.chat.ChatThreadEditScreen
+import com.github.familyvault.ui.components.filesCabinet.DocumentUploadActionButton
+import com.github.familyvault.ui.components.filesCabinet.ImageUploadActionButton
 import com.github.familyvault.ui.screens.main.tasks.task.TaskNewScreen
 import com.github.familyvault.ui.theme.AppTheme
 import familyvault.composeapp.generated.resources.Res
 import familyvault.composeapp.generated.resources.chat_create_new
-import familyvault.composeapp.generated.resources.file_cabinet_sending_files
-import familyvault.composeapp.generated.resources.file_cabinet_upload
 import familyvault.composeapp.generated.resources.task_new
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.IO
-import kotlinx.coroutines.withContext
 import org.jetbrains.compose.resources.stringResource
 import org.koin.compose.koinInject
 
@@ -104,112 +88,11 @@ class MainScreen : Screen {
 
     @Composable
     private fun FloatingFileCabinetActionButton() {
-        val imagePicker = koinInject<IImagePickerService>()
-        val documentPicker = koinInject<IDocumentPickerService>()
-        val fileCabinetService = koinInject<IFileCabinetService>()
         val currentTabIndex = FilesCabinetTab.selectedTabIndex
 
-        var startImagePicker by remember { mutableStateOf(false) }
-        var startDocumentPicker by remember { mutableStateOf(false) }
-        var isUploading by remember { mutableStateOf(false) }
-        var errorMessage by remember { mutableStateOf<String?>(null) }
-
         when (currentTabIndex) {
-            0 -> {
-                FloatingActionButton(onClick = {
-                    startImagePicker = true
-                }) {
-                    Icon(
-                        Icons.Filled.UploadFile,
-                        stringResource(Res.string.file_cabinet_upload),
-                    )
-                }
-
-                if (startImagePicker) {
-                    LaunchedEffect(Unit) {
-                        val images = imagePicker.pickImagesAndReturnByteArrays()
-
-                        if (images.isNotEmpty()) {
-                            isUploading = true
-
-                            withContext(Dispatchers.IO) {
-                                images.forEach { mediaByteArray ->
-                                    fileCabinetService.sendImageToFamilyGroupStore(mediaByteArray)
-                                }
-                            }
-
-                            isUploading = false
-                        }
-
-                        startImagePicker = false
-                    }
-                }
-            }
-            1 -> {
-                FloatingActionButton(onClick = {
-                    startDocumentPicker = true
-                }) {
-                    Icon(
-                        Icons.Filled.UploadFile, 
-                        stringResource(Res.string.file_cabinet_upload),
-                    )
-                }
-
-                if (startDocumentPicker) {
-                    LaunchedEffect(Unit) {
-                        try {
-                            val documents = documentPicker.pickDocumentsAndReturnByteArrays()
-
-                            if (documents.isNotEmpty()) {
-                                isUploading = true
-                                
-                                withContext(Dispatchers.IO) {
-                                    val documentUrls = documentPicker.getSelectedDocumentUrls()
-
-                                    documentUrls.forEachIndexed { index, uri ->
-                                        val documentName = documentPicker.getDocumentNameFromUri(uri) ?: "document_$index.pdf"
-                                        val documentMimeType = documentPicker.getDocumentMimeTypeFromUri(uri) ?: "application/pdf"
-
-                                        documents.getOrNull(index)?.let { docBytes ->
-                                            fileCabinetService.sendDocumentToFamilyGroupStore(
-                                                docBytes,
-                                                documentName,
-                                                documentMimeType
-                                            )
-                                        }
-                                    }
-                                }
-                                
-                                isUploading = false
-                            } else {
-                            }
-                        } catch (e: Exception) {
-                            errorMessage = "Error: ${e.message}"
-                        } finally {
-                            startDocumentPicker = false
-                        }
-                    }
-                }
-            }
-        }
-
-        if (isUploading) {
-            CircularProgressIndicatorDialog(stringResource(Res.string.file_cabinet_sending_files))
-        }
-        
-        errorMessage?.let { error ->
-            AlertDialog(
-                onDismissRequest = { errorMessage = null },
-                title = { Text("Error") },
-                text = { Text(error) },
-                confirmButton = {
-                    TextButton(
-                        onClick = { errorMessage = null }
-                    ) {
-                        Text("OK")
-                    }
-                }
-            )
+            0 -> ImageUploadActionButton()
+            1 -> DocumentUploadActionButton()
         }
     }
